dpam/requirements.txt
# Core dependencies
psycopg2-binary>=2.9.5
requests>=2.28.1
numpy>=1.23.4
gemmi>=0.5.5

# Web and API
fastapi>=0.89.1
uvicorn>=0.20.0
pydantic>=1.10.4
python-multipart>=0.0.5

# Visualization
matplotlib>=3.6.2
seaborn>=0.12.1

# Utilities
tqdm>=4.64.1
pandas>=1.5.2
biopython>=1.80
jsonschema>=4.17.3

# Testing
pytest>=7.2.0
pytest-cov>=4.0.0

# Documentation
Sphinx>=5.3.0
sphinx-rtd-theme>=1.1.1

# Development
black>=22.12.0
isort>=5.11.4
flake8>=6.0.0
mypy>=0.991